[Invibes] Adapter rework with review changes
