+++ invibes/invibes-htb-exports.js
@@ -1,0 +1,5 @@
+//? if (FEATURES.GPT_LINE_ITEMS) {
+shellInterface.InvibesHtb = {
+    render: SpaceCamp.services.RenderService.renderDfpAd.bind(null, 'InvibesHtb')
+};
+//? }
